+++ example_bookticker.py
@@ -1,0 +1,64 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# File: example_bookticker.py
+#
+# Part of ‘UNICORN Binance WebSocket API’
+# Project website: https://github.com/unicorn-data-analysis/unicorn-binance-websocket-api
+# Documentation: https://www.unicorn-data.com/unicorn-binance-websocket-api.html
+# PyPI: https://pypi.org/project/unicorn-binance-websocket-api/
+#
+# Author: UNICORN Data Analysis
+#         https://www.unicorn-data.com/
+#
+# Copyright (c) 2019, UNICORN Data Analysis
+# All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a
+# copy of this software and associated documentation files (the
+# "Software"), to deal in the Software without restriction, including
+# without limitation the rights to use, copy, modify, merge, publish, dis-
+# tribute, sublicense, and/or sell copies of the Software, and to permit
+# persons to whom the Software is furnished to do so, subject to the fol-
+# lowing conditions:
+#
+# The above copyright notice and this permission notice shall be included
+# in all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+# OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABIL-
+# ITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT
+# SHALL THE AUTHOR BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+# WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+# IN THE SOFTWARE.
+
+from unicorn_binance_websocket_api.unicorn_binance_websocket_api_manager import BinanceWebSocketApiManager
+import time
+import threading
+
+
+def print_stream_data_from_stream_buffer(binance_websocket_api_manager):
+    while True:
+        if binance_websocket_api_manager.is_manager_stopping():
+            exit(0)
+        oldest_stream_data_from_stream_buffer = binance_websocket_api_manager.pop_stream_data_from_stream_buffer()
+        if oldest_stream_data_from_stream_buffer is False:
+            time.sleep(0.01)
+        else:
+            print(oldest_stream_data_from_stream_buffer)
+
+
+# create instance of BinanceWebSocketApiManager and provide the function for stream processing
+binance_websocket_api_manager = BinanceWebSocketApiManager(exchange="binance.com-futures")
+
+# start one worker process (or more) to move the received stream_data from the stream_buffer to a print function
+worker_thread = threading.Thread(target=print_stream_data_from_stream_buffer, args=(binance_websocket_api_manager,))
+worker_thread.start()
+
+# create streams
+bookTicker_arr_stream_id = binance_websocket_api_manager.create_stream("arr", "!bookTicker")
+
+while True:
+    binance_websocket_api_manager.print_stream_info(bookTicker_arr_stream_id)
+    time.sleep(1)